Added example of how Google Analytics can be used
Extended Backbone.View with a function to raise the ga:trackEvent global
event, passing in the tracking parameters as an object literal.

The main.js also listens for the same event, passing in the values to
Google Analytics.

The SessionView.js raises the event using the new trackEvent helper
function, passing in all four parameters:

category: Session
Action: favorite
Label: The ID of the session
Value:  1 or -1.

By passing in a value of 1 or -1, we can use the Google Analytic event
trackign to calculate the average value for a given Action + Label, and
that way figure out how many people have favorited the session.